Switch agentId from usize to u32

+++ npc-engine-core/src/lib.rs
@@ -15,7 +15,7 @@ pub use util::*;
 
 #[derive(Copy, Clone, Debug, Hash, PartialEq, Eq, PartialOrd, Ord, serde::Serialize)]
 #[serde(rename_all = "kebab-case")]
-pub struct AgentId(pub usize);
+pub struct AgentId(pub u32);
 
 // TODO: remove debug constraints
 pub trait NpcEngine: Sized + 'static {
